Fixed inifinite while loop bug with helm chart tgz parser
By default, the tgz parser expects the first item in the tgz
is going to be the "module"/Chart.yaml file. This may not be
enforced by all the tools who might generate a chart tgz file.

The fix is to continue the while loop and iterate to the next
items in the tar to find the Chart.yaml file.